fix: update contact only if fields are not empty

+++ src/js/helpers/editingContacts.js
@@ -20,27 +20,40 @@ export const editingContacts = () => {
         
         inputName.focus()
       } else {
-        if (confirm('Сохранить изменения?')) {
-          inputName.disabled = true
-          inputPhone.disabled = true
-          parentNode.classList.remove('_edit')
+        if (inputName.value.length > 0 && inputPhone.value.length > 0) {
+          if (confirm('Сохранить изменения?')) {
+            inputName.disabled = true
+            inputPhone.disabled = true
+            parentNode.classList.remove('_edit')
+  
+            const id = inputId.value
+            const group = inputGroup.value
+            const name = inputName.value
+            const phone = inputPhone.value
+  
+            const contactsList = JSON.parse(localStorage.getItem('contactsBook'))
+  
+            contactsList[group].map((item) => {
+              if (item.userId === id) {
+                item.userName = name
+                item.userPhone = phone
+              }
+              return item
+            })
+  
+            localStorage.setItem('contactsBook', JSON.stringify(contactsList))
 
-          const id = inputId.value
-          const group = inputGroup.value
-          const name = inputName.value
-          const phone = inputPhone.value
-
-          const contactsList = JSON.parse(localStorage.getItem('contactsBook'))
-
-          contactsList[group].map((item) => {
-            if (item.userId === id) {
-              item.userName = name
-              item.userPhone = phone
+            if (inputName.classList.contains('_error')) {
+              inputName.classList.remove('_error')
             }
-            return item
-          })
-
-          localStorage.setItem('contactsBook', JSON.stringify(contactsList))
+            if (inputPhone.classList.contains('_error')) {
+              inputPhone.classList.remove('_error')
+            }
+          }
+        } else {
+          !inputName.value.length && inputName.classList.add('_error')
+          !inputPhone.value.length && inputPhone.classList.add('_error')
+          alert('Вы оставили пустые поля, так нельзя!')
         }
       }
     }
